move group cli tests to e2e tests
