+++ tests/e2e/group/query.go
@@ -1,4 +1,4 @@
-package testutil
+package group
 
 import (
 	"fmt"
