+++ tests/e2e/group/tx.go
@@ -1,4 +1,4 @@
-package testutil
+package group
 
 import (
 	"encoding/base64"
